extract combined_perf_score.ts from metrics.ts with CPS config and normalization functions
- update imports across Svelte components and tests to new CPS module
- remove redundant CPS configuration from `metrics.ts`
